requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
qrcode==7.4.2
Pillow==10.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
